feat(trustroots):refactor-common-components
