+++ client/src/app/shared-components/StatusIcon.tsx
